Split check_X_Y_domain and check_X_domain to more functions
