+++ skada/_utils.py
@@ -6,7 +6,6 @@
 
 import logging
 from numbers import Real
-from typing import Optional, Set
 
 import numpy as np
 
@@ -16,7 +15,6 @@ from sklearn.covariance import (
     ledoit_wolf,
     shrunk_covariance,
 )
-from sklearn.utils import check_array, check_consistent_length
 from sklearn.utils.multiclass import type_of_target
 
 _logger = logging.getLogger('skada')
@@ -46,239 +44,6 @@ def _estimate_covariance(X, shrinkage):
     elif isinstance(shrinkage, Real):
         s = shrunk_covariance(empirical_covariance(X), shrinkage)
     return s
-
-
-def check_X_y_domain(
-    X,
-    y,
-    sample_domain=None,
-    allow_source: bool = True,
-    allow_multi_source: bool = True,
-    allow_target: bool = True,
-    allow_multi_target: bool = True,
-    return_indices: bool = False,
-    # xxx(okachaiev): most likely this needs to be removed as it doesn't fit new API
-    return_joint: bool = False,
-    allow_auto_sample_domain: bool = True,
-    allow_nd: bool = False,
-):
-    """
-    Input validation for domain adaptation (DA) estimator.
-    If we work in single-source and single target mode, return source and target
-    separately to avoid additional scan for 'sample_domain' array.
-
-    Parameters:
-    ----------
-    X : array-like of shape (n_samples, n_features)
-        Input features
-    y : array-like of shape (n_samples,)
-        Target variable
-    sample_domain : array-like or None, optional (default=None)
-        Array specifying the domain labels for each sample.
-    allow_source : bool, optional (default=True)
-        Allow the presence of source domains.
-    allow_multi_source : bool, optional (default=True)
-        Allow multiple source domains.
-    allow_target : bool, optional (default=True)
-        Allow the presence of target domains.
-    allow_multi_target : bool, optional (default=True)
-        Allow multiple target domains.
-    return_indices : bool, optional (default=False)
-        If True, return only the source indices.
-    return_joint : bool, optional (default=False)
-        If True, return X, y, and sample_domain without separation.
-    allow_auto_sample_domain : bool, optional (default=True)
-        Allow automatic generation of sample_domain if not provided.
-    allow_nd : bool, optional (default=False)
-        Allow X and y to be N-dimensional arrays.
-
-    Returns:
-    ----------
-    If return_indices is True :
-        source_idx : array
-            Boolean array indicating source indices.
-    If return_joint is False :
-        X_source : array
-            Input features for source domains.
-        y_source : array
-            Target variable for source domains.
-        X_target : array
-            Input features for target domains.
-        y_target : array
-            Target variable for target domains.
-    If return_joint is True :
-        X : array
-            Input features
-        y : array
-            Target variable
-        sample_domain : array
-            Array specifying the domain labels for each sample.
-    """
-
-    X = check_array(X, input_name='X', allow_nd=allow_nd)
-    y = check_array(y, force_all_finite=True, ensure_2d=False, input_name='y')
-    check_consistent_length(X, y)
-    if sample_domain is None and not allow_auto_sample_domain:
-        raise ValueError("Either 'sample_domain' or 'allow_auto_sample_domain' "
-                         "should be set")
-    elif sample_domain is None and allow_auto_sample_domain:
-        y_type = _check_y_masking(y)
-        sample_domain = _DEFAULT_SOURCE_DOMAIN_LABEL*np.ones_like(y)
-        # labels masked with -1 (for classification) are recognized as targets,
-        # labels masked with nan (for regression) are recognized as targets,
-        # the rest is treated as a source
-        if y_type == 'classification':
-            mask = (y == _DEFAULT_MASKED_TARGET_CLASSIFICATION_LABEL)
-        else:
-            mask = (np.isnan(y))
-        sample_domain[mask] = _DEFAULT_TARGET_DOMAIN_LABEL
-    else:
-        sample_domain = check_array(
-            sample_domain,
-            dtype=np.int32,
-            ensure_2d=False,
-            input_name='sample_domain'
-        )
-        check_consistent_length(X, sample_domain)
-
-    source_idx = (sample_domain >= 0)
-    # xxx(okachaiev): this needs to be re-written to accommodate for a
-    # a new domain labeling convention without "intersections"
-    n_sources = np.unique(sample_domain[source_idx]).shape[0]
-    n_targets = np.unique(sample_domain[~source_idx]).shape[0]
-    if not allow_source and n_sources > 0:
-        raise ValueError(f"Number of sources provided is {n_sources} "
-                         "and 'allow_source' is set to False")
-    if not allow_target and n_targets > 0:
-        raise ValueError(f"Number of targets provided is {n_targets} "
-                         "and 'allow_target' is set to False")
-    if not allow_multi_source and n_sources > 1:
-        raise ValueError(f"Number of sources provided is {n_sources} "
-                         "and 'allow_multi_source' is set to False")
-    if not allow_multi_target and n_sources > 1:
-        raise ValueError(f"Number of targets provided is {n_targets} "
-                         "and 'allow_multi_target' is set to False")
-
-    if return_indices:
-        # only source indices are given, target indices are ~source_idx
-        return source_idx
-    elif not return_joint:
-        # commonly used X, y, X_target, y_target format
-        return X[source_idx], y[source_idx], X[~source_idx], y[~source_idx]
-    else:
-        return X, y, sample_domain
-
-
-# xxx(okachaiev): code duplication, just for testing
-def check_X_domain(
-    X,
-    sample_domain,
-    *,
-    allow_domains: Optional[Set[int]] = None,
-    allow_source: bool = True,
-    allow_multi_source: bool = True,
-    allow_target: bool = True,
-    allow_multi_target: bool = True,
-    return_indices: bool = False,
-    # xxx(okachaiev): most likely this needs to be removed as it doesn't fit new API
-    return_joint: bool = True,
-    allow_auto_sample_domain: bool = True,
-):
-    """
-    Input validation for domain adaptation (DA) estimator.
-    If we work in single-source and single target mode, return source and target
-    separately to avoid additional scan for 'sample_domain' array.
-
-    Parameters:
-    ----------
-    X : array-like of shape (n_samples, n_features)
-        Input features.
-    sample_domain : array-like of shape (n_samples,)
-        Domain labels for each sample.
-    allow_domains : set of int, optional (default=None)
-        Set of allowed domain labels. If provided, only these domain labels are allowed.
-    allow_source : bool, optional (default=True)
-        Allow the presence of source domains.
-    allow_multi_source : bool, optional (default=True)
-        Allow multiple source domains.
-    allow_target : bool, optional (default=True)
-        Allow the presence of target domains.
-    allow_multi_target : bool, optional (default=True)
-        Allow multiple target domains.
-    return_indices : bool, optional (default=False)
-        If True, return only the source indices.
-    return_joint : bool, optional (default=True)
-        If True, return X and sample_domain without separation.
-    allow_auto_sample_domain : bool, optional (default=True)
-        Allow automatic generation of sample_domain if not provided.
-
-    Returns:
-    ----------
-    If return_indices is True:
-        source_idx : array
-            Boolean array indicating source indices.
-    If return_joint is False:
-        X_source : array
-            Input features for source domains.
-        X_target : array
-            Input features for target domains.
-    If return_joint is True:
-        X : array
-            Combined input features for source and target domains.
-        sample_domain : array
-            Combined domain labels for source and target domains.
-    """
-    X = check_array(X, input_name='X')
-    if sample_domain is None and not allow_auto_sample_domain:
-        raise ValueError("Either 'sample_domain' or 'allow_auto_sample_domain' "
-                         "should be set")
-    elif sample_domain is None and allow_auto_sample_domain:
-        # default target domain when sample_domain is not given
-        # xxx(okachaiev): I guess this should be -inf instead of a number
-        # The idea is that with no labels we always assume
-        # target domain (_DEFAULT_TARGET_DOMAIN_ONLY_LABEL)
-        sample_domain = (
-            _DEFAULT_TARGET_DOMAIN_ONLY_LABEL * np.ones(X.shape[0], dtype=np.int32)
-        )
-    else:
-        sample_domain = check_array(
-            sample_domain,
-            dtype=np.int32,
-            ensure_2d=False,
-            input_name='sample_domain'
-        )
-        check_consistent_length(X, sample_domain)
-    if allow_domains is not None:
-        for domain in np.unique(sample_domain):
-            # xxx(okachaiev): re-definition of the wildcards
-            wildcard = np.inf if domain >= 0 else -np.inf
-            if domain not in allow_domains and wildcard not in allow_domains:
-                raise ValueError(f"Unknown domain label '{domain}' given")
-
-    source_idx = (sample_domain >= 0)
-    n_sources = np.unique(sample_domain[source_idx]).shape[0]
-    n_targets = np.unique(sample_domain[~source_idx]).shape[0]
-    if not allow_source and n_sources > 0:
-        raise ValueError(f"Number of sources provided is {n_sources} "
-                         "and 'allow_source' is set to False")
-    if not allow_target and n_targets > 0:
-        raise ValueError(f"Number of targets provided is {n_targets} "
-                         "and 'allow_target' is set to False")
-    if not allow_multi_source and n_sources > 1:
-        raise ValueError(f"Number of sources provided is {n_sources} "
-                         "and 'allow_multi_source' is set to False")
-    if not allow_multi_target and n_sources > 1:
-        raise ValueError(f"Number of targets provided is {n_targets} "
-                         "and 'allow_multi_target' is set to False")
-
-    if return_indices:
-        # only source indices are given, target indices are ~source_idx
-        return source_idx
-    elif not return_joint:
-        # commonly used X, y, X_target, y_target format
-        return X[source_idx], X[~source_idx]
-    else:
-        return X, sample_domain
 
 
 def _merge_source_target(X_source, X_target, sample_domain) -> np.ndarray:
